move FileOps into vcc/utils move logger into vcc/utils
