+++ Audio.cpp
@@ -24,7 +24,7 @@ This file is part of VCC (Virtual Color Computer).
 #include "config.h"
 #include "coco3.h"
 #include "audio.h"
-#include <vcc/common/logger.h>
+#include <vcc/utils/logger.h>
 #include "BuildConfig.h"
 #include "Cassette.h"
