+++ Cassette.cpp
@@ -25,7 +25,7 @@ This file is part of VCC (Virtual Color Computer).
 #include <vcc/common/DialogOps.h>
 #include "Cassette.h"
 #include "stdio.h"
-#include <vcc/common/logger.h>
+#include <vcc/utils/logger.h>
 #include <functional>
 
 unsigned char MotorState=0,TapeMode=STOP,WriteProtect=0,Quiet=30;
