+++ FD502/fd502.cpp
@@ -23,9 +23,9 @@
 #include "utils.h"
 #include "../CartridgeMenu.h"
 #include <vcc/devices/serial/beckerport.h>
-#include <vcc/common/FileOps.h>
 #include <vcc/common/DialogOps.h>
-#include <vcc/common/logger.h>
+#include <vcc/utils/FileOps.h>
+#include <vcc/utils/logger.h>
 #include <vcc/utils/winapi.h>
 #include <vcc/devices/rtc/oki_m6242b.h>
 #include <vcc/devices/rom/rom_image.h>
