+++ Vcc.cpp
@@ -42,7 +42,7 @@ This file is part of VCC (Virtual Color Computer).
 #include <commdlg.h>
 #include <stdio.h>
 #include <mmsystem.h>
-#include <vcc/common/FileOps.h>
+#include <vcc/utils/FileOps.h>
 #include <vcc/common/DialogOps.h>
 #include "defines.h"
 #include "resource.h"
@@ -66,7 +66,7 @@ This file is part of VCC (Virtual Color Computer).
 #include "Cassette.h"
 
 #include "CommandLine.h"
-#include <vcc/common/logger.h>
+#include <vcc/utils/logger.h>
 #include "memdump.h"
 
 #include "MemoryMap.h"
