+++ keyboardEdit.cpp
@@ -36,7 +36,7 @@ This file is part of VCC (Virtual Color Computer)
 #include "defines.h"
 #include "config.h"
 
-#include <vcc/common/logger.h>
+#include <vcc/utils/logger.h>
 #include "keyboard.h"
 #include "keyboardLayout.h"
 #include "keyboardEdit.h"
