+++ libcommon/include/vcc/utils/dll_deleter.h
@@ -16,7 +16,7 @@
 //	VCC (Virtual Color Computer). If not, see <http://www.gnu.org/licenses/>.
 ////////////////////////////////////////////////////////////////////////////////
 #pragma once
-#include <vcc/common/logger.h>
+#include <vcc/utils/logger.h>
 #include <Windows.h>
