+++ libcommon/src/utils/Fileops.cpp
@@ -18,7 +18,7 @@
 #include <Windows.h>
 #include <stdio.h>
 #include <stdbool.h>
-#include <vcc/common/FileOps.h>
+#include <vcc/utils/FileOps.h>
 
 void ValidatePath(char *Path)
 {
