+++ orch90/orchestra90cc_cartridge.cpp
@@ -17,7 +17,7 @@
 ////////////////////////////////////////////////////////////////////////////////
 #include "orchestra90cc_cartridge.h"
 #include "resource.h"
-#include <vcc/common/FileOps.h>
+#include <vcc/utils/FileOps.h>
 #include <vcc/utils/winapi.h>
 #include <vcc/utils/filesystem.h>
 #include <Windows.h>
